working inputs
