+++ .stickler.yml
@@ -1,13 +1,15 @@
-# add the linters you want stickler to use for this project
 linters:
+  rubocop:
+    display_cop_names: true
+  csslint:
+  stylelint:
   eslint:
-    # indicate where is the config file for stylelint
-    config: './.eslintrc.json'
-
-# add the files here you want to be ignored by stylelint
+    config: './eslint.config'  
 files:
   ignore:
-      - 'dist/*'
+    - 'dist/*'
+    - '.eslintrc.json'
+    - 'node_modules/*'
 
 # PLEASE DO NOT enable auto fixing options
 # if you need extra support from you linter - do it in your local env as described in README for this config
