+++ src/DOM/projectInput.js
@@ -1,8 +1,6 @@
 import { addProjectToList } from '../control/projectControl';
 
-const newProjectInput = () => {
+export default function newProjectInput() {
   const projectInput = document.getElementById('projectInput').value;
   addProjectToList(projectInput);
-};
-
-module.exports = { newProjectInput, default: newProjectInput };
+}
