+++ src/control/projectControl.js
@@ -1,5 +1,4 @@
-/* eslint-disable no-undef */
-import addProjectToTaskDropdown from '../DOM/projectDOM';
+import { addProjectToTaskDropdown } from '../DOM/projectDOM';
 import { projectList, Project } from '../modules/constructor';
 import { updateLocalStorage } from '../modules/localStorage';
 import { BulmaModal } from '../modules/bulmaModal';
@@ -14,6 +13,7 @@ export const renderProjects = () => {
 export const addProjectToList = (projectInput) => {
   const projectModal = new BulmaModal('#projectModal');
   if (projectList.some((project) => project.name === projectInput)) {
+    // eslint-disable-next-line no-undef
     swal('Oops', 'Duplicate project!', 'error');
   } else {
     const id = projectList.length;
